Abort completion when opening parentheses
Often cmp will compete a function name and (by design) not insert the
opening parenthesis. However, if we then enter this opening parentheses,
this does not dismiss the completion window. The result is that if we
then hit tab again, we will accept the completion, removing the
parenthesis we added. The fix is just to abort completions when typing
"(", which seems to be a special case.

We also fix backspace behaviour, where the completion would be
"remembered" erroneously.