Tweak completion behaviour
Replacing cmp.abort with cmp.close on backspace fixes a bug, where the
abort will restore the line to its initial, pre-completion state which
is probably not what we want when we hit backspace to the beginning of
the line.

The tab behaviour change makes it so we always complete to the longest
common substring, rather than waiting for another tab.

Other changes are potential performance improvements, where we avoid
repeated assignments inside function calls and so on.