Solve Git completion issue in zsh
Git completion has been not optimal on zsh on macOS for me since around
the end of 2023 or start of 2024. It seems like this was caused either
by installing Git from Homebrew or changing the way I installed
Homebrew, as the root cause is that I was using the Git-provided zsh
completions instead of the zsh-native ones. This was happening because
$fpath's first entry was /opt/homebrew/share/zsh/site-functions, set by
brew shellenv in ~/.zprofile.

We take the first element of $fpath and put it at the end, which solves
the problem nicely (assuming the first entry is always Homebrew's).

Solution from https://github.com/orgs/Homebrew/discussions/2797.