Improve inlay hint display
Previous inlay hints were coloured gray, to make them look similar to
comments, and to disambiguate they weren't italicised.

This made them stand out too much mid-code, as they were too dark.

We bump the brightness and italicise, the convention being that italics
signifies that it's not code, or something like that. A darker version
of the highlight colour for types is an obvious choice, but it's too
noisy and also hints aren't just for types but include struct fields and
so o, so we go with a gray-ish blue.