Rearrange conditions on startx alias
Use -z instead of negation to appease shellcheck.

We swap the order because it's likely faster to check if a variable is
set than a command exists, but also because on systems with startx
present we have to do more checks most of the time (since the startx
test would pass, so we would continue on to check $DISPLAY).