Follow nvim-treesitter rewrite
nvim-treesitter has swapped the default branch and, as of 2025-12-14,
considers main to be stable.

The main loss is peeking at definition, as we no longer have LSP
interoperability. We also lose native support for automatic installation
and enabling tree-sitter by default, but both can be worked around (the
former by using lewis6991/ts-install.nvim, the latter with an
autocommand).

Overall the configuration is less elegant, but functional.