Skip running golangci-lint-ls on go.mod files
When running golangci-lint in a directory that does not contain Go files
(such as a module that contains packages in subdirectories only), we get
the error

    % golangci-lint run .
    ERRO Running error: context loading failed: failed to load packages:
    failed to load packages: package .: no go files to analyze

which won't be changed by golangci-lint
(https://github.com/golangci/golangci-lint/issues/6092). This produces
spurious diagnostics when run from the language server in such cases,
which calls golangci-lint with the current directory.

One possible fix would be to ignore these errors in the language server.
This is done in
https://github.com/nametake/golangci-lint-langserver/pull/55, but
there's no release yet.

To be reverted once
https://github.com/nametake/golangci-lint-langserver/issues/64 is closed
and a new release is picked up by package managers.